Make Embedding integral part of GeneTreeInSpeciesNetwork
In my mind, Embedding does not make any sense when not given a target (i.e. a
SpeciesNetwork) and the separation into two levels of wrappers was confusing.

Therefore I suggest that one level of wrapping is exactly what is needed, and
now GeneTreeInSpeciesNetwork does the heavy lifting also of tracking Embeddings.
However, this is not tested yet.